+++ mapcache-ios/ui/GPKGSLoadTilesViewController.m
@@ -11,6 +11,7 @@
 #import "GPKGSUtils.h"
 #import "GPKGSProperties.h"
 #import "GPKGSConstants.h"
+#import "GPKGSDecimalValidator.h"
 
 NSString * const GPKGS_LOAD_TILES_SEG_GENERATE_TILES = @"generateTiles";
 
@@ -18,6 +19,7 @@ NSString * const GPKGS_LOAD_TILES_SEG_GENERATE_TILES = @"generateTiles";
 
 @property (nonatomic, strong) NSArray * urls;
 @property (nonatomic, strong) GPKGSGenerateTilesViewController *generateTilesViewController;
+@property (nonatomic, strong) GPKGSDecimalValidator * epsgValidator;
 
 @end
 
@@ -30,13 +32,18 @@ NSString * const GPKGS_LOAD_TILES_SEG_GENERATE_TILES = @"generateTiles";
 
     self.urls = [GPKGSProperties getArrayOfProperty:GPKGS_PROP_PRELOADED_TILE_URLS];
     
+    self.epsgValidator = [[GPKGSDecimalValidator alloc] initWithMinimumInt:-1 andMaximumInt:99999];
+    [self.epsgTextField setDelegate:self.epsgValidator];
+    
     UIToolbar *keyboardToolbar = [GPKGSUtils buildKeyboardDoneToolbarWithTarget:self andAction:@selector(doneButtonPressed)];
     
     self.urlTextField.inputAccessoryView = keyboardToolbar;
+    self.epsgTextField.inputAccessoryView = keyboardToolbar;
 }
 
 - (void) doneButtonPressed {
     [self.urlTextField resignFirstResponder];
+    [self.epsgTextField resignFirstResponder];
 }
 
 -(void)alertView:(UIAlertView *)alertView clickedButtonAtIndex:(NSInteger)buttonIndex {
@@ -53,9 +60,13 @@ NSString * const GPKGS_LOAD_TILES_SEG_GENERATE_TILES = @"generateTiles";
                     NSNumber * maxZoom =[url objectForKey:GPKGS_PROP_PRELOADED_TILE_URLS_MAX_ZOOM];
                     NSNumber * defaultMinZoom =[url objectForKey:GPKGS_PROP_PRELOADED_TILE_URLS_DEFAULT_MIN_ZOOM];
                     NSNumber * defaultMaxZoom =[url objectForKey:GPKGS_PROP_PRELOADED_TILE_URLS_DEFAULT_MAX_ZOOM];
+                    NSNumber * epsg = [url objectForKey:GPKGS_PROP_PRELOADED_TILE_URLS_EPSG];
                     
                     [self.urlTextField setText:urlValue];
                     self.data.url = self.urlTextField.text;
+                    
+                    [self.epsgTextField setText:[epsg stringValue]];
+                    self.data.epsg = [epsg intValue];
                     
                     [self.generateTilesViewController setAllowedZoomRangeWithMin:[minZoom intValue] andMax:[maxZoom intValue]];
                     
@@ -102,6 +113,10 @@ NSString * const GPKGS_LOAD_TILES_SEG_GENERATE_TILES = @"generateTiles";
 
 - (IBAction)urlChanged:(id)sender {
     self.data.url = self.urlTextField.text;
+}
+
+- (IBAction)epsgChanged:(id)sender {
+    self.data.epsg = [self.epsgTextField.text intValue];
 }
 
 -(void)prepareForSegue:(UIStoryboardSegue *)segue sender:(id)sender{